requirements/testing.txt
nose2-cov
coveralls>=0.2